+++ src/features/navigation/NavBar.tsx
@@ -105,27 +105,7 @@ export function NavBar() {
                     </Link>
                   )}
                 </Menu.Item>
-                <Menu.Item>
-                  {({ active }) => (
-                    <Link
-                      href={NAV_ITEMS.collectionBuilds.href}
-                      className={cn(
-                        active ? 'bg-gray-800' : '',
-                        'flex w-full flex-row items-start justify-start p-2 text-sm font-semibold text-white',
-                      )}
-                    >
-                      <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.collectionBuilds.icon className="h-5 w-5 text-primary-600" />
-                      </div>
-                      <div className="flex flex-col items-start justify-start gap-y-1">
-                        {NAV_ITEMS.collectionBuilds.label}
-                        <p className="text-xs font-normal text-gray-400">
-                          {NAV_ITEMS.collectionBuilds.description}
-                        </p>
-                      </div>
-                    </Link>
-                  )}
-                </Menu.Item>
+
                 <Menu.Item>
                   {({ active }) => (
                     <Link
@@ -440,23 +420,6 @@ export function NavBar() {
 
                     <p className="text-xs text-gray-400">
                       {NAV_ITEMS.communityBuilds.description}
-                    </p>
-                  </div>
-                </Link>
-
-                <Link
-                  href={NAV_ITEMS.collectionBuilds.href}
-                  className="flex flex-row items-center justify-start"
-                >
-                  <NAV_ITEMS.collectionBuilds.icon
-                    className="mr-2 h-7 w-5 flex-none text-primary-600"
-                    aria-hidden="true"
-                  />
-                  <div className="flex flex-col items-start justify-start px-3 py-2">
-                    {NAV_ITEMS.collectionBuilds.label}
-
-                    <p className="text-xs text-gray-400">
-                      {NAV_ITEMS.collectionBuilds.description}
                     </p>
                   </div>
                 </Link>